Add files via upload
This is file is a C++ source code for projecting points on line networks that are stored in QGIS as polylines. The program takes the geometrical definition of point data and polyline data as input and projects the points on the network. For every point, the nearest point on the network is computed and is coupled with the original point. This utility can be used in instances where polyline data (such as road or pipeline network) and point data (such as bus stops or standposts) is obtained from online sources such as OSM, OpenLayers or municipal databases - that have significant offsets and measurement errors.